+++ CHANGES.md
@@ -65,3 +65,5 @@
 
 ### 12/27/24
 - Created get user data which uses a users sessions ID as a token to get user data
+- Fix profile where it shows your actual username
+